+++ lib/screens/result.dart
@@ -4,12 +4,51 @@ class Result extends StatelessWidget {
   const Result({required this.shape, required this.results, super.key});
 
   final String shape;
-  final List<double> results;
+  final Map<String, dynamic> results;
+
+List<String> readKeys() {
+  List<String> keys = [];
+  for (var key in results.keys) {
+    keys.add(key);
+  }
+  return keys;
+}
+
+List<double> readValues() {
+  List<double> values = [];
+  for (var value in results.values) {
+    values.add(value);
+  }
+  return values;
+}
 
   @override
   Widget build(BuildContext context) {
-    return const Center(
-      child: Text('Results are in!'),
+    return Scaffold(
+      body: SafeArea(
+        child: Column(
+          children: [
+            Center(
+              child: Text(
+                'Results for $shape',
+                // style: TextStyle(...),
+              ),
+            ),
+            ListView.builder(
+              itemCount: results.length,
+              shrinkWrap: true,
+              itemBuilder: (context, index) {
+                return Row(
+                  children: [
+                    Text(readKeys()[index]),
+                    Text(readValues()[index].toStringAsFixed(2)),
+                  ],
+                );
+              },
+            ),
+          ],
+        ),
+      ),
     );
   }
 }